+++ packages/spacecat-shared-utils/src/index.js
@@ -41,10 +41,15 @@ export {
   getQuery,
 } from './helpers.js';
 
+export {
+  isAWSLambda,
+} from './runtimes.js';
+
 export { sqsWrapper } from './sqs.js';
 export { sqsEventAdapter } from './sqs.js';
 
 export { logWrapper } from './log-wrapper.js';
+export { instrumentAWSClient } from './xray.js';
 
 export {
   composeBaseURL,