+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,31 @@
+# [@adobe/spacecat-shared-utils-v1.37.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.37.1...@adobe/spacecat-shared-utils-v1.37.2) (2025-05-06)
+
+
+### Bug Fixes
+
+* **xray:** local friendly xray instrumentation ([#714](https://github.com/adobe/spacecat-shared/issues/714)) ([7c7cebd](https://github.com/adobe/spacecat-shared/commit/7c7cebdde4e1c0b24c80201c5d16a68f470122cb))
+
+# [@adobe/spacecat-shared-utils-v1.37.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.37.0...@adobe/spacecat-shared-utils-v1.37.1) (2025-05-03)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#712](https://github.com/adobe/spacecat-shared/issues/712)) ([76cebd2](https://github.com/adobe/spacecat-shared/commit/76cebd2a7a7b9799e4ca265833620eada01f5c8c))
+
+# [@adobe/spacecat-shared-utils-v1.37.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.36.3...@adobe/spacecat-shared-utils-v1.37.0) (2025-04-28)
+
+
+### Features
+
+* **xray:** non-aws ie local friendly xray instrumentation ([#710](https://github.com/adobe/spacecat-shared/issues/710)) ([5184b4d](https://github.com/adobe/spacecat-shared/commit/5184b4dc3ddff8a6e44d12129475fa323533be2c))
+
+# [@adobe/spacecat-shared-utils-v1.36.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.36.2...@adobe/spacecat-shared-utils-v1.36.3) (2025-04-26)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#708](https://github.com/adobe/spacecat-shared/issues/708)) ([57535ac](https://github.com/adobe/spacecat-shared/commit/57535ac8c636de229cec7c5cee83dead07ac09fb))
+
 # [@adobe/spacecat-shared-utils-v1.36.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.36.1...@adobe/spacecat-shared-utils-v1.36.2) (2025-04-15)